Extract account mapping hooks
